+++ tinygrad/features/graph/cuda.py
