+++ tinygrad/runtime/ops_webgpu.py
@@ -1,9 +1,7 @@
-import functools
 from wgpu.utils.device import get_default_device
 from tinygrad.device import Compiled, Allocator
 from tinygrad.codegen.kernel import LinearizerOptions
-from tinygrad.renderer.cstyle import uops_to_cstyle
-from tinygrad.renderer.wgsl import WGSLLanguage
+from tinygrad.renderer.cstyle import WGSLRenderer
 import wgpu
 
 wgpu_device = get_default_device()
@@ -34,4 +32,4 @@ class WebGpuAllocator(Allocator):
 class WebGpuDevice(Compiled):
   def __init__(self, device:str):
     super().__init__(WebGpuAllocator(), LinearizerOptions(device="WEBGPU", supports_float4=False, local_max=[256, 256, 64], global_max=[65535, 65535, 65535]),
-                     functools.partial(uops_to_cstyle, WGSLLanguage()), lambda x: x, WebGPUProgram)
+                     WGSLRenderer, lambda x: x, WebGPUProgram)
